requirements.txt
django>=1,<4
djangorestframework>=3,<4
psycopg2>=2.7.1,<2.9  # https://stackoverflow.com/questions/68024060/assertionerror-database-connection-isnt-set-to-utc
aiohttp>=3.5.0,<4.0.0
azure-storage-blob  # todo: find adapted requirements
applicationinsights>=0.11.2,<0.12.0